Move header files to a more appropriate home
