+++ opencog/atoms/pattern/PatternTerm.h
